+++ src/doctestfilters.jl
@@ -21,6 +21,7 @@ GAP_docs_pages = [
         "conversion.md",
         "packages.md",
         "other.md",
+        "signals.md",
         "examples.md",
         "internal.md",
         "manualindex.md",